xks/requirements.txt
asgiref==3.2.7
Django==2.2.11
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter==2.2.0
djangorestframework==3.11.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
